allow failed job be resumed
